fix: move TestClock to time module as requested
- Move TestClock from concurrency to time.testing package
- Place TestClock in time module main source for cross-module access
- Update concurrency tests to import TestClock from time module
- Add coroutines-test as implementation dependency in time module
- Remove old TestClock from concurrency module test directory
- All tests passing with proper TestClock dependency injection
