+++ modules/lib/concurrency/src/test/java/com/duchastel/simon/photocategorizer/concurrency/RateLimiterImplTest.kt
@@ -1,5 +1,6 @@
 package com.duchastel.simon.photocategorizer.concurrency
 
+import com.duchastel.simon.photocategorizer.time.testing.TestClock
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.async
 import kotlinx.coroutines.delay
