+++ modules/lib/time/src/main/java/com/duchastel/simon/photocategorizer/time/testing/TestClock.kt
@@ -1,11 +1,12 @@
-package com.duchastel.simon.photocategorizer.concurrency
+package com.duchastel.simon.photocategorizer.time.testing
 
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineScheduler
 import kotlin.time.Clock
 import kotlin.time.ExperimentalTime
 import kotlin.time.Instant
 
-@OptIn(ExperimentalTime::class)
+@OptIn(ExperimentalTime::class, ExperimentalCoroutinesApi::class)
 class TestClock(
     private val scheduler: TestCoroutineScheduler
 ) : Clock {
